Add api.ClientCreate method
